Refactor verifyAccess and its tests
Wrote test to take read access into account and had to change
performSortingTest in the ChronicleModelTest file to take a path
parameter as a string. Also added unit test for the setErrorMessage
method. Also fixed so the middle panel will color the folder path output
to red when a error is being displayed.