Remove getters from controller and update codebase
Removed the getters from the ChronicleController and updated testfiles
that used them. Also added a runOnEDT helper method in the controller...
as well as added some methods to verify the main frame shows when the
showMainFrame method is called, updated the
shouldPassIfDialogOpensUpInCenterOfApp method to be more reliable, added
methods to ensure the info dialog displays the correct message when the
sortFolder method is called with valid and invalid paths and that the
path label in the middle panel gets colored red when an error is
displayed for the folder.