Change error message passed into displayError
Changed it to something more general that will fit neatly in the
JDialog. Also changed error to throw in the verifyAccess method for null
paths to a checked exception so the user will be notified if ever that
would happen. It uses the GENERAL_ERROR from AppConfig, and I noticed a
dot was missing at the end of the sentence, so I added that as well.
Also had to update the shouldThrowForNullFolder and
shouldHandleEmptyPathInSortFolderMethod test methods for the model to
take theses changes into account.